chore: make dtype configurable

+++ src/langchain_openvino/utils.py
@@ -274,7 +274,7 @@ def validate_parameters(
     )
 
 
-def decrypt_model(model_dir, model_file_name, weights_file_name):
+def decrypt_model(model_dir, model_file_name, weights_file_name, dtype=np.uint8):
     """
     Loads and (optionally) decrypts the OpenVINO model and weights files.
 
@@ -293,7 +293,7 @@ def decrypt_model(model_dir, model_file_name, weights_file_name):
     with open(model_dir + "/" + weights_file_name, "rb") as file:
         binary_data = file.read()
     # decrypt weights
-    weights = np.frombuffer(binary_data, dtype=np.uint8).astype(np.uint8)
+    weights = np.frombuffer(binary_data, dtype=dtype).astype(dtype)
 
     return model, Tensor(weights)
